Revised Operation Torch. Here's how it works:
- Operation Torch can be performed by the USA about 2 months after Barbarossa (Should be in August, 1941?)
- If Torch is not performed by February 2, 1942, the German Reich annexes Vichy.
- Until Torch or the German vichy annexation, Vichy is entirely neutral

This needs some more revisions, though. I'd say this is pretty barebones right now; future plans include a fleet OOB to be transferred to Germany or Italy, a temporary naval invasion speed buff to the Allies, and the possibility for Tunisia to be transferred to Italy at the French armistice in 1939 if Italy is being bold and joins the war before France is dying.

Like I said: for now, it'll do.