Reworked GBP.
First off, the removal of Staff Office plan will help this doctrine assume a unique identity and playstyle.

GBP is now the thinking man's choice for slow, methodological warfare. The distinguishing factor of the GBP tree is its planning bonus. 30% before it splits, 30% after on each end.

The left side, Assault, is focused on tanks and motorized. It's a bit of a contradiction; tanks and motorized that plan slowly and hit hard when they finally do. I suspect this branch is going to struggle to find a niche at first, but I'd rather make it underpowered than OP. It should not be beating out Mobile Warfare or Deep Battle as a choice, for most countries.

The right side, Infiltration, is focused on an infantry assault on foot, preferably along a wide but static front. Raj and Japan might be interested in this.

Organization for this doctrine is pretty high. Across the board, it has some of the highest starting organization--The catch is that your recovery rate is pretty questionable.

With enough time to prepare, GBP should really, really hurt. However, it should also be vulnerable during this preparation period in some respects and struggle in continual combat when planning bonus is worn down.

The hope is that all this is going to make GBP feel different from other doctrines, we'll see.