We'll see where this goes, but I've noticed that UK's and everyone, really, is spending a lot of IC on planes. This goes for Romania, Australia, and USA, too--LOTS of people that have a designated 'plane factory' that just produces planes for them.
This is a sign that the cost of aircraft is too high in general, to me. That, and the fact that the UK is commonly getting stretched REALLY thin and even getting totally blasted in the air war, what's up with that? They shouldn't have more planes than Germany, sure, but they shouldn't be losing in the one zone they choose to fight in, either.

Thusly, I've reduced the cost of planes from 2x vanilla to 1.5x the cost, this is along with the airbase change. We'll see where this goes, hope to see IC being spent on other things aside from "aircraft x1000000". There really needs to be a point at which planes stop being the premium thing you need to spam more of.