2 infra and 1 AA added to Marsa Matruh. Historical? Nah. Probably not. Chalk it up as desert fortifications leading up to Alamain or something.
What this is meant to do is encourage the UK to not just camp on Alamain and encourage them to go aggress. In a similar way, it's meant to make it easier for the Axis to aggress in this direction. Camp on Alamain? Okay, Italy's going to swipe Marsa and build on it easier with that AA. We'll see where this goes.