air support base down to 30, 35 was way too high
Light Tank 3 costs 1 oil now

The purge rebalance is done. The Soviets suffer badly at first, but will recover just in time for the Defense of Moscow and will come back even stronger in 1942.

The Soviets now have an infantry specialist.

The Soviets get a 50% Medium 3 bonus on 'The Red Army Resurgent'.

Germany now has a 50% Panther bonus on Battle of Raisenthai. Techs are not date-locked yet, but will be soon to a reasonable extent.

Stalin is no longer a coward with Quick Surrender, but Stalingrad and Baku are worth more VP's. Better defend them.

Soviets now start with CAS 1 and AA 1, and get a 60% CAS 2 research bonus.

The work on the Soviet Union, as of this commit, is finished. Most of it. There will surely be more.

Death to the German invader!