- Changing armor on Heavy Tanks so they shouldn't get pierced by out of date TD's, but I haven't done proper testing on this.
- Several spammy French events have been consolidated into one that summarizes the situation after France falls and removes all the bad spirits
- Some MTTH have been purged
- Events giving research bonuses to Germany removed, these are in the focus tree now

And now, the big finale of this commit:

The *Second Armistice at Compiègne rework*.

The Second Armistice at Copiegne and establishment of the Vichy State has been reworked. It is no longer event-based, because the event caused lag and was also a gigantic eyesore. In addition, exploitative uses such as Germany holding the event and such are no longer possible.

The exact way this works is a little complex, but to summarize: Vichy is now established the exact moment France capitulates. The exact borders of Vichy vary a bit based primarily on how much territory Italy has taken from it.

If Italy is greedy and joins the war against France and successfully makes gains, they may keep certain territories they gain. These territories include Provence, Savoy, Alpes, Corsica, Tunisia, and Gabes.

But joining comes with a risk, as it will remove the Phoney War from the allies and leave Italy vulnerable to attack in Africa. Joining the war as Italy before Germany is done with France has pros and cons. You'll strengthen your position in the Meditarranean and get some more industry, but can you hold out until Germany arrives to help you?