Error log now at about 3,500~. It was at 5k yesterday.
The bad news: A lot of the errors (Think 3,000) is coming from History/States, because the map was changed around to make room for important, vital tags in Africa. Removing the entire folder will reset all industry, because that's where factories in states are stored.

Thank you, Paradox. This is gonna take a bit of work to fix.