New process for Operation Countenance and conflict over Iraq.
In short: Once Germany and Italy are at war with the UK, the UK can invade Iraq. Once Iraq capitulates, they annex it.

If Alexandria is threatened (UK does not fully control) OR the date is greater than 1940 June, the UK can perform Operation Countenance, which annexes Iran and allows them to set up a defense there. The Raj must be in control of Baluschistan and the UK must own Baghdad to do this. The Soviets get the territory they'd normally get from Countenance. Balance adjustments will be done to account for this situation.

If the UK does NOT invade Iraq, then Germany can do it once they are in control of Jordan.

If the UK does not invade Iran, then Germany can do it once they are in control of Baghdad.

In short: The battle of Africa continues, even after Alamain falls. The allies need to be prepared to fight a defensive war in Iran, if it comes down to it... Which isn't too hard, let's be completely honest. It's mountainous and terrible.