Norway and Denmark now function similarly to the Benelux, in that Germany must now declare war on both countries simultaneously.
These changes should hope give the feeling of the 'lightning war' to consoldiate Europe. Complacency is no longer an option; move to consolidate quickly or the UK will beat you to the punch.

German-Hungarian partition of Czechoslovakia restored as an option

The German annexation of Austria and Czechoslovakia not reducing MEFO bill cost fixed. I didn't know this was even a thing, it explains why I was PP starved that one game I played Germany. Oof.