Boy, oh boy, here we go.
Welcome to the China Rework commit... Part one.

Changes:

USA:

- Can support China with a decision to spend civs to give them some offmap mils.
- Can support China by sending 250 fighters through the Flying Tigers

JAPAN:
- Receives 50 air and army XP for every state in China they take from China (may be buggy, need to test but can't because game keeps crashing due to map). This basically means China isn't incentivzed to avoid fighting Japan, or discouraged from putting up air
- Given 4 civs and 3 mils to compensate for the monster I've possibly unleashed.

CHINA:

- Modified WTT tree & content restored. Several useless focuses that did nothing have been taken out, several focuses cut in half time-wise.
- Industrial evacuations readded
- French Cooperation, American Cooperation, British Cooperation are now only options in foreign investors tree
- British and French cooperation are no longer mutually exclusive
- Focuses underneath 'Invite Foreign Investors' no longer need relation improvements
- 'Invite the Flying Tigers' allows the USA to send 250 fighters to China.
- Chinese mil focuses changed to be useful, achievable without crashing economy
- Rural Reconstruction Movement buffed
- United Front triggers when Japan begins Marco Polo Bridge, China annexes warlords. All of them.
- China gets a modified 'Lure them in Strategy' from PFU, but slightly less obnoxious. Maybe.
- China starts with a line on artillery and AA
- Welfare tree not cursed. Each focus can be used as an emergency inflation reduction button. This should probably be looked at because it doesn't make sense, but this was the best solution I had to make the choice meaningful
- Building slots added in distant backwoods so factories can relocate somewhere
- P L A Y A B L E (pending map changes.)

COMMUNIST CHINA:

- Given shock therapy designed to eliminate AIDS
- Playable

NOW INTRODUCING:

The FIXED and BETTER BURMA ROAD!

After China completes the 'Burma Road' focus and has control of Yunnan (After Japan attacks), decisions open up to invest in the Burma Road.

Investing in the Burma Road as the UK or China costs 4 civs for 60 days, along with 25 PP.

After you invest in the Burma Road, China gets 3 off-map Military Factories.

However, if China loses the southwestern part of its country (Or the Raj does, technically), then the Burma Road closes and you lose all your invested mils.

There's a cooldown period of 90 days.

The UK now has an opportunity to fight Japan early and indirectly using China as a proxy, but it comes at the cost of some economy. It's an option, though. Japan is going to want to go to Burma to shut this down... Which requires investment, because of serious supply and infrastructure problems.

This is actually apparently somewhere close to how the Burma Road was supposed to work, by the way, except Paradox broke it by not coding it properly.

And no, just to clarify, my goal isn't to make an omega China that kills Japan; that's not really the intention. China is supposed to get pushed back, lose a lot of ground, and not pose an actual threat unless allowed to fester for several years after that point. Y'know, due to the burma road.