SPECIAL FORCES UPDATE:
China:
- Some more inflation added. Sorry, not sorry.

SPECIAL FORCES are now reworked.

- Special forces take less time to research
- Special Forces no longer use a percentage of regiments modifier; instead, they work off of a flat cap.
- The Special Forces technology is cheaper and allows you to get 4~ more special forces max.
- Amphibious Tanks and Mechanized no longer count as Special Forces.

The default cap regiment-wise is 160. All majors get a spirit that raises this by 320. The UK, USA, and Japan get a super-spirit available through their focus tree that raises it by 385 and for Japan 400.

Special Forces are now rare and very powerful instead of swarming entire islands. This is also not a system that can be gamed or exploited; your cap is your cap. No more 500+ unequipped divisions to get marines.

Generally speaking:

- Minors can have 11 Special Forces 40w 14/4, or 10 Special Forces with AT and AA (2 of each and then art)
- Majors can have roughly 34~ 14/4 special forces, or 30 Special Forces with AT and AA
- The UK and USA can have roughly 38~ 14/4 special forces, or 34 Special Forces with AT and AA
- Japan can have 40~ 14/4 special forces, or 35 special forces with AT and AA