France, Case Anton, and Vichy updated from a naval perspective.
France now begins with its full historical fleet. This could lead to some exploitative and highly undesired situations, so a rule will be implemented that says France cannot put its fleet in the Med AND cannot delete it. Defending Norway is fair game, though.

After the capitulation of France and the Armistice, the entirety of the French fleet is transferred to the Vichy state.

The UK can perform the Mers-El-Kebir Raid in response, blowing up 2 early Vichy battleships, 2 cruisers, and 6 destroyers.

When the time comes to invade Vichy and perform Torch, the Axis responds by performing Case Anton. If the Allies wait too long (March 1942), the Axis can perform Case Anton regardless as a punishment for the Allies stalling.

Unlike in real life, though, Case Anton is a success and the Vichy fleet, what remains of it, ends up in Axis hands. Germany gets a choice: Do they keep the fleet for themselves, or let Italy have it?

This series of events is a bit more historical, and also more interesting. While it may be 'too little, too late', the introduction of a hundred or so screens and some subs will serve as a much needed injection into the Axis' lategame naval power, allowing them to claw their way back in to naval relevancy in the Meditarranean under decent air cover. Refitting and upgrading is a lot cheaper than in vanilla, too, so that should help you.