AIRBASE LIMITER ACTIVITING.
The airbase limiter is a scripted thing inside the game that uses an on_action and some scripted variables to limit the size of airbases in certain locations!

The Pacific is meant for island hopping and carrier warfare. Large islands, like Borneo and the Philippines, have a size of 3; small Pacific islands have a size of one.

With this change, covering yourself purely with land-based air isn't an option anymore. You'll need to use escort carriers and island hopping in conjunction to take control of territory.

This, of course, again, will be paired with buffs to land based aircraft accordingly, to make the Med more about land based VS Pacific which is heavily carrier based.

Happy to say this system is coded very well with few lag problems, thanks to some help of some folks over at the Modding Co-op