Rubber nerfs.
Less stuff in Singapore, less rubber in Ceylon, no rubber in Tamil, less rubber in China, rubber in Bismarck.

Nerf to USA's rubber company.

Look, I'm going to be honest. I just absolutely massacred rubber production in Ceylon--You know why? Because it was stupid. Why? Because if you have half a brain, you can just defend it and be fine forever. You build up the infrastructure to 10, now you have 200 rubber + excvacation tech. That's dumb. The allies shouldn't be surviving off of just Ceylon alone; it means they ignore Japan and play way too defensively.

However, there's some more rubber in New Guinea, which should make it more of a battleground. You need to be fighting for all your rubber so Japan doesn't take it, and if he does? You need to take it back or your plane production gets shut down.

More adjustments will be needed but this is a start.