Re-initializing Airbase Limiter.
Does it lag? Why, no. It doesn't.

Why?

Because it uses an A R R A Y, BABY.

What's an array? Think of it as a digital storage box for things in a program.

Instead of looping through every state on the planet for every person every day, and then determining if the state matches the criteria to have an airbase removed, the two Arrays simply loops through themselves every day (Only once per day!) to check a pre-determined list of state ID's.

This massive list is compiled only once on startup.

If it has a state id that is flagged as being small, it removes an airbase if it is above 1. If it is medium, it removes it if it i above 3.

Praise be to the modding co-op community.