Gently trimming one civilian factory off the UK.
Why? Because people are going to figure out the meta soon, and it's not going to be pretty.

What do I mean by that?

The UK gets total mobilization in 1939 pretty much every game. It's the allied leader and has decent everything; the rational strategy is trading with the UK. People haven't always been doing that because they haven't figured it out, but they're wisening up.

This, on top of the USA "buffs", means that the UK could end up really REALLY horrifyingly strong. This is preventative. We'll see how it works.