Big fixes to map, strategic regions, etc.
Less cancerous airzones in early Russia for Germany allowing them to be competitive with the red air force.

State added in Khartoum (Port Sudan) for more airports and easier pushing in Egypt.

Lower Egypt is more of a challenge to push, but maybe that's fine. we'll just need to see. I feel like Italy is one grindy breakthrough away from getting killed in this area so it's a risky proposition to camp.