Implementation of the 'real Battle of France'.
So, how does this work?

First off, an introduction:

One of the goals of GDU is giving each country the feeling of having autonomy and the ability to do things that most players would like to do. Most players want to play HOI; this typically involves building units, an army, and doing things with these units.

Whenever the ideal gameplay does not coincide with this style, it's a sign that something is wrong and in need of a fix.

As of late, the perceived ideal playstyle for France is a plane-focused style where France puts everything onto building fighters and ships the motherload to the UK. This strategy was already nerfed, but the main problem with it is that it was still too viable and there was no real reason not to do it.

Upon Germany invading Belgium, France gets a mission for 35 days to 'organize the resistance'--In short, if you are capitulated before these 35 days have passed, you don't get to access any of your Free France focuses, losing out on research bonuses, about 25 dockyards and factories for the rest of the game, and just generally being a worthless shell compared to what could've been. The thing is, this shouldn't be happening assuming a half-competent player is on France that tries to actually hold. This provides the allies with some incentive to fight in France, but provides dimishing returns after the timer has run out.

In addition, Germany gets a 5% stability buff which is just enough to catapult him over 90% stability, not only boosting production but giving him -1% consumer good--Which is potentially significant considering the amount of industry GER controls.

Furthermore, there's plans to attack the plane spam problem at the source by limiting (Not removing!) the impact of the infamous gangbang effect, allowing some CAS to join battles even if air is red, and nerfing air in general.

In addition, to prevent the offloading of the job of plane spammer to a Commonwealth member, all their air bonuses have basically been cut in half.

Some more changes to France are intended to make them more fun to play beyond this, including new land advisors after their demise, a new focus tree (courtesy of Revan, inspected by myself, it's actually not that bad for a first tree and after some touching up it's quite good and even pleasant to look at, credit to Revan here), and the ability to actually fight Germany for a while even if the air is questionable.

This update will be 1.5, when it is ready for release... Otherwise known as the 'French update'.