Concealment Expert for Russia going bye-bye because it was contributing to the possibility to go no-planes.
Replaced with Infantry Specialist and Commando Specialist. This would be a problem, but let me just remind you that you're not allowed to mix Special Forces and Infantry in the same unit. I guess the Soviets can use this instead of Roko early Barb if they're going for an infantry focused strategy and Caucasus is not a relevant theatre.

INDUSTRY_AIR_DAMAGE_FACTOR (AA) down to 0.12, but it gets better from AA technology. Up to 90% reduction, pretty much like vanilla, which Paradox did OK for once.