Because of the previous changes to the defines, CAS is now harder to intercept so there's less of a harsh barrier of "you have planes!" and "you don't have planes"; it's more of a gradual process.
Unfortunately, this *also* means that bombers are a bit harder to intercept and will often get through to some extent.

As a result, tactical and strategic bombers are getting a nerf. They should trade pretty badly if not escorted. I'll keep an eye on this and make more changes as needed.