So, this commit is gonna be a bit of a doozy.
In short: Trying to resolve the 'motorized' VS tank variant problem.

Motorized/truck-drawn AT, AA, and Artillery is simply too good right now for how much it costs. The solution for this is basically just siding with the tanks on this one and saying that they're meant to have better stats, generally speaking. That said, the truck-drawn stuff is OK. It isn't terrible, but you probably want tank variants. I think that's how it should be. Otherwise, we end up in a meta where truck drawn AT reigns supreme along with a single TD for piercing.

In short, towed motorized/AT/AA is now treated as a direct competitor with SPG/TD/SPAA and balanced as such. May need fine tuning.

Supply zones should be less horrifying to push into but there's a bit less supply to compensate

MAJOR changes to air defines--These changes make the air war less snowbally and allow for CAS to fly even at low air superiority, though plenty is intercepted.

Fine-tuning supply on infantry and artillery