Trying something.
In short: removing Germany's 5% attack factor on Short War because it was a nasty interaction in some cases and stacks hard. Multiplicative modifiers are very scary and it's time to de-inflate the mod where possible. This is contributing to Germany feeling unstoppable, having the best tanks, generals, doctrine, and CAS is more than enough. Instead, there's a 2.5% recovery rate attached (Maybe this is too little, but Germany already gets lots of recovery from General Staff so I wanna be careful)

Total War is mostly untouched and has 5% recovery and 5% attack

This will help the allies feel like they have a more fair shot fighting the Germans, but some adjustments in other areas will be needed.