Retuning some of the Soviet Union bonuses.
Basically, the -10% production cost on Planes/Tanks, while cool, had a lot of problems.

You need both of those things as the Soviets, yes, but having it exclusively for those things encouraged a lot of tank and plane spam, which isn't ideal. Furthermore, it can make it hard to replenish your gun stockpile--Which is kinda the thing the Soviets need.

The solution here is a quick adjustment of their Production Targets spirit. This commit is removing their -10% production cost on tanks/planes and instead giving them +10% production efficiency max (So, +20% total) on Production Targets.

This is, obviously, a net Soviet buff--More production efficiency helps a lot, especially from day one. The design goal of the Soviets is to not give them an equipment problem, but a purge problem; I'm considering other options to this end, but I'm also afraid of doing too much too fast.

Some other options on the table to get the Soviets to the playstate where I want them to be:

- Supply consumption buff to go along with additional fielded equipment
- Making purge slightly worse to compensate for large amounts of extra gear on field.