Second pass on Acclimation changes. I've got a general idea of how this works, but we'll need to see in game.
In short:

- Base Acclimation is very low and difficult to get, but fairly rewarding once you actually get it.
- Axis mainland has a hard time with Acclimation in Winter and will not be able to cheesily farm it. Penalty goes away when at war with Soviets.
- Soviets and Finland get massive amounts of Acclimation to where they are the main benefactors of it.
- General traits relating to Acclimation swapped around because otherwise they'd be very mandatory with these changes

The idea here is introducing Acclimation as a variable that can help the Soviets really turn dangerous during Winter, opening up opportunities to allow Germany to punch harder in 1942 and Barbarossa while giving the Soviets a window to do their thing.

Need to see this in action, but from what I can tell it's working as designed and at this point it's a matter of number tweaking. Germany really won't get a lot of acclimation throughout the game, I'd expect... 40% at most if they really work for it. This is subject to change; Soviets are supposed to have an advantage in Winter but if it's too much, it can be dialed back. Just need to experiment.