this should actually fix france and also the special forces problem on the techs, if not I dunno what I'm gonna do
Vichy France now forms in bite-sized pieces to reduce lag and destroy ANY possibilities that this was somehow related to that...

I put the Vichy event and on_action in its own file to avoid ANY contamination or risks

it DOES NOT CRASH right now in single player. if it crashes in MP, welp, I dunno. I even tested the saves that were sent to me, neither crashed. If things aren't working at this point, I really don't know what is wrong with it