Carefully tweaking a lot of details about the map, supply regions, and airfields in Africa and to a lesser extent the Middle East. The objective is to encourage a longer Africa campaign with more defensible locations and more areas similar to Marsa Matruh/Cyrenaica where more tiles are added to allow for mobile warfare. Depending on how people play, if you lose 1 area, you can probably still be able to defend another area rather than having to entirely abandon Africa. Of course, this will only be the case if Axis chooses to actually fight there and also to use air there, but they are incentivized to do so. Regarding airfield locations, note that efficiency is calculated from the center of a state rather than the location of the airfield, so airfield location within a state only influences how hard it is for one side or another to take an airfield.
Generally:

1) Tobruk, Benghasi, Tripoli, and Tunisia should all be to a significant extent defensible on their own to a greater extent even if other cities/ports have been lost. If Axis loses eastern Libya, that doesn't necessarily mean that you have to wholesale abandon the entirety of Africa any more.
2) The allies should still have a tough time defending Egypt against a committed Axis, but have a small supply advantage (which probably doesn't matter at El Alemein, where quality > quantity). For the moment no major change in how defensible Egypt is is intended, but that could potentially be adjusted in the future.
3) When the Allies Torch, Axis should have a tough time trying to fight in Morocco/western Algeria (limited supply, but an easier time fighting in Tunisia and eastern Algeria, where they should have some ability to defend and counterattack). The intent is to allow more of a possibility of a Tunisia campaign/Kasserine pass.
4) Allies/Soviets have some additional defensive advantages from airfields/supply in Iran and Sudan.
5) Supply is limited (but present) in Africa, hopefully this means the Allies have enough troops to fight on in other areas even if they get some wiped in one area. Hopefully this also means that Axis can continue to attack/defend/contest in Africa for a long time without feeling like they are pulling away too many resources from Barbarossa. May need further adjustment one way or the other after further testing.

Specifics:
A) Commonwealth should have the same supply as UK in Africa, unless UK builds up infra/ports beyond its focus (commonwealth port levels buffed to make that so).
B) Axis supply will come ultimately from Tripoli rather than Benghasi, after Italy does its Libyan Infrastructure focus.
C) Oil/rubber is moved around a bit to adjust player incentives. Some oil is moved to Marsah Matruh (and a small amount to Cyrenaica) to further encourage active fighting in these areas, but UK will have to ask itself if it really wants to build up infrastructure too much in Marsah Matruh and possibly give Axis more oil. Other oil/rubber is moved from Benghasi/Sirte to Tripoli, so Axis has more incentive to defend that as well even if eastern Libya is lost (this is partly also done for some technical reasons relating to incentives for the supply system).
D) Sea tiles north of Tobruk and Tunisia moved to the Central Med and Tyrrhenian Sea to make them more defensible for Axis.
E) Shapes of some tiles in Marsa Matruh/Cyrenaica tweaked.
F) Axis should have a significant supply and air advantage in eastern Libya that the Allies will have to fight hard to overcome. Airfields in eastern Libya are also moved to mountain tiles close to/on ports, and Benghasi and the tile to its south is mountain rather than a hill.
G) Axis should also have a significant supply advantage and possibly even an air zone advantage in western/central Libya. Many additional tiles are also added in this area to allow for more mobile warfare. So even if eastern Libya falls, Rommel has reason to fight on from Tripoli.
H) It should be fairly hard to attack Tunisia from either Libya or Algeria due to supply advantages for the defenders in Tunisia, though significantly less so if Allies can attack from both Libya and Algeria. An additional state is added in Tunisia which provides another airfield. If Allies take Tunisia, this also means more air over Sicily and Italy later on though. A few extra mobile warfare tiles added.
I) 2 desert tiles added in southern Algeria near Morocco, so there is no continuous mountain line (helps ensure Allies can attack into the Algeria air zone if Axis has managed to get to the mountain line).
J) Attacking into Egypt, Allies should still have a significant supply advantage even after El Alemein falls, and some additional tiles to the east of El Alemein are added for mobile warfare. This may make it possible for Allies to fight/retreat in good order at least a bit better in Egypt after El Alemein falls.
K) Sinai is moved from the Egypt air zone to the Levant air zone. By itself this makes the Egypt air zone a bit less favorable for the Allies, but this is somewhat counterbalanced by states in Libya shifting their centers to the west. This also makes Suez a bit more defensible for the allies if Egypt has fallen, since it is not in the Egypt air zone.
L) In Jordan, a desert tile is added to the south of the mountain line, so that at least a small amount of mobile defense is required rather than entrenching into mountains (defending from either side).
M) The Mesopotamia air zone becomes a bit more favorable to Axis defenders defending against Allies attacking from Iran. That is, if the Axis air controller doesn't send all the planes to the Soviet Union... Deir ez-Zor is moved from the Levant to the Mesopotamia air zone and many airfields in the Mesopotamia air zone are moved to the west.
N) A new supply zone is added in the deserts between Iraq and Syria/Jordan. Supply is limited here, and extra tiles are added for mobile warfare. This might be somewhat defensible for the Allies due to Axis supply issues, as long as the Allies have enough tanks. Later on, it may be fairly defensible for the Axis, particularly if they contest air and fight a mobile defense with counterattacks from the baseline of the Jordan mountain line.
O) Some airfields in Iran are shifted to a bit more defensible positions (especially in northern Iran, where they are shifted back to defensible mountains for the Soviets). Axis supply should be pretty limited in Iran unless they make great efforts to improve supply.
P) Axis should have some supply limitations attacking south into Sudan from Egypt. Meanwhile, Allied supply is a bit buffed in Sudan, and airfields in Sudan are moved to defensible mountains on/near the port of Port Sudan.
Q) It is not possible to build infrastructure in the impassable portions of the Sahara desert.

R) Unrelated to Africa - some airfields far back in the Soviet Union are moved behind the Volga (last Soviet line of defense). In the Raj, the Dacca airfield is moved behind the Dacca line.