Rework of Japan's army/navy rivalry to encourage balance between some army and some navy rather than autistic singular focus on only army (or only navy, but more often it seems to be only army).
Factory/Dockyard output is replaced with Production Cost reductions because factory output affects air as well as army equipment (including carrier planes and NAVs), and because both navy and army focused builds need air (particularly some fighters).

Having balance between the 2 gives you generally the best bonuses, so probably you want to take 2 of the army options and 2 of the navy options most of the time.