Air losses redone in light of some testing and the successful validation of an air losses model. This is designed to work IN CONJUNCTION WITH reverting back to the previous plane costs (which were 1.5x vanilla plane cost).
The goal of reducing plane spam by enabling the side that is losing the air war (producing fewer planes) to nonetheless keep reasonably contesting the air in a good # of air zones for a longer period of time should be achieved, while keeping planes more expensive so they are less tedious to micro.

With 1941 industry and 1940 fighters, each factory that the side which is producing fewer planes (usually the Axis) has on planes increases the equilibrium long-run size of its air force by about 70 planes (more with 1943 industry or with the Soviet production targets), assuming that you constantly are fighting with all of your planes and are trading 1:1. However, the losing side's air force should not be ground down to lower than about 70 planes per factory that you have producing at full efficiency. For example, if you have 150 factories on planes, even if you are losing the air war and constantly using your planes, your air force shouldn't drop below about 10500 planes (give or take), which provides you with a reasonable number of planes to keep the air contested to a reasonable degree.

A big part of the problem was that how bloody the air war was varied a lot because of upgrades and modifiers (from aces and veterancy) to fighters' air attack, and the # of planes shot down varies greatly with variations in air attack. This meant that in particular in Barbarossa, when planes were kept up for a long time and all the air wings became veteran and got aces, and when often both sides had enough XP to upgrade fighter guns to some degree, the air war became extremely deadly, so that the losing side's plane count shrunk very rapidly and precipitously. To avoid this, aces and veterancy of air wings no longer affect air attack, but instead affect agility and speed to slightly greater degrees. You also can no longer upgrade fighter air attack, but only agility and speed (separately, rather than combined in "engine" upgrades).

The define for how many planes are shot down in an individual battle (COMBAT_DAMAGE_SCALE) is also lowered significantly. While this does mean that there are fewer losses in any **individual day** of air combat, this doesn't mean that losses no longer matter or that there are fewer losses over the long-run of sustained air combat (i.e. Barbarossa and continuous use of air in Africa). Actually there are not generally fewer losses overall because air losses depend to a large degree (linearly) on the overall number of planes that are fighting. For the side that is losing the air war, assuming all planes are being used, daily losses will asymptotically approach daily production of the side that is producing fewer planes, while the other side will build up an increasing stockpile of (eventually unused) planes.

Finally, there are some QOL improvements to make microing air a bit less painful, including faster deployment of planes (1 hour deployment time) and a maximum of 6400 planes which can be deployed at once (which you can then split into wings of 100 evenly).