--- The Chairman Trump Memorial France Update ---
France is re-worked to further give France players the right incentives - ***to fight with honor in France*** - and to render impossible exploits such as saving all the French equipment in the training queue when France capitulates and then immediately naval invading Norway with 30 French 40-w marine divisions in 1939, *not that anyone would ever dream of doing such a thing*...

When France capitulates, technically it becomes a different country with a different tag (FRF rather than FRA), which means it is not possible to directly save equipment, but the player is automatically switched to this "new" Free France country with a new focus tree.

Instead of France being entirely locked out of its Free France focus tree if it doesn't hold 30 days, France and Germany are put in competition for 10 Civs and 10 Mils. For every week that France holds, up to a maximum of 10 weeks, it will get 1 civ and 1 mil in its Free France focus tree. All the civs and mils that France does *not* get instead immediately go to Germany as off-map factories. For example, if France holds for 30 days (more than 4 weeks), it will get 4 civs and 4 mils, and Germany will get the remaining 6 civs and 6 mils. And if France holds for 8 more than weeks, France will get 8 civs and 8 mils while Germany will get 2 civs and 2 mils, etc. France also gets --- or doesn't get --- some additional civs/mils/docks from its Free France focus tree depending on how long it holds. Since France is going to lose all its equipment, it may as well initially focus on properly trying to defend France, holding on as long as possible, and inflicting as much damage as possible on Germany. Meanwhile Germany has every reason to try to kill France as quickly as possible.

If France doesn't hold for at least 7 days though, they are still locked out entirely of their Free France focus tree (surely it should be impossible to non-intentionally capitulate that fast?)

When the player is automatically switched to the "new" Free France country, they should be able to 1-day research their old techs, will start with a lot of PP and XP to re-pick advisors, and start with an equipment stockpile for about 5 divisions (1 light armor). Effectively these are the divisions that are "saved" from mainland France. They have temporary national spirits allowing them to train their early divisions very quickly and to reach maximum production efficiency on their starting factories within a couple of days, so they effectively don't lose out on their efficiency from the country change.

This is partly also intended to give Germany a bit of extra industry to be able to sustain losses in Africa (if France lasts as long as it normally did previously, this is probably worth about an extra 5 or so civs/mils for Germany). It also discourages meme German navy builds etc by giving Germany a greater incentive to focus on killing France quickly so as to get more industry and deny more industry to Free France.

Still needs checking on some things and working out a few additional details, but just committing partial (mostly complete) work to get it on the git.