1) All UK/USA/Commonwealth land doctrine research bonuses reduced from 100% to 50%.
2) USA starts with an "inexperienced army" national spirit that gives -30% land doctrine research speed; this is removed from the Louisiana Maneuvers focus.

The overall intent of these changes is that the Allies shouldn't have their land doctrine finished super-early, and also that they should be a bit less tempted by mobile warfare. Although they can still go mobile warfare, the idea is that if they do, they shouldn't reap the full benefits until something like 1941/42 rather than 1939/1940 (in Africa etc). This may not necessarily be enough to fully achieve this, but is an initial step.